+++ src/command_modules/azure-cli-storage/azure/cli/command_modules/storage/_validators.py
@@ -16,7 +16,7 @@ from azure.cli.core.commands.client_factory import get_mgmt_service_client
 from azure.cli.core.commands.validators import validate_key_value_pairs
 
 from ._factory import get_storage_data_service_client
-from .util import glob_files_locally
+from .util import glob_files_locally, guess_content_type
 
 storage_account_key_options = {'primary': 'key1', 'secondary': 'key2'}
 
@@ -78,6 +78,11 @@ def validate_accept(namespace):
             'full': TablePayloadFormat.JSON_FULL_METADATA
         }
         namespace.accept = formats[namespace.accept.lower()]
+
+
+def validate_bypass(namespace):
+    if namespace.bypass:
+        namespace.bypass = ', '.join(namespace.bypass) if isinstance(namespace.bypass, list) else namespace.bypass
 
 
 def validate_client_parameters(namespace):
@@ -299,7 +304,7 @@ def validate_blob_type(namespace):
         namespace.blob_type = 'page' if namespace.file_path.endswith('.vhd') else 'block'
 
 
-def get_content_setting_validator(settings_class, update):
+def get_content_setting_validator(settings_class, update, guess_from_file=None):
     def _class_name(class_type):
         return class_type.__module__ + "." + class_type.__class__.__name__
 
@@ -359,9 +364,11 @@ def get_content_setting_validator(settings_class, update):
             new_props.content_language = new_props.content_language or props.content_language
             new_props.content_md5 = new_props.content_md5 or props.content_md5
             new_props.cache_control = new_props.cache_control or props.cache_control
+        else:
+            if guess_from_file:
+                new_props = guess_content_type(ns[guess_from_file], new_props, settings_class)
 
         ns['content_settings'] = new_props
-        namespace = argparse.Namespace(**ns)
 
     return validator
 
@@ -780,6 +787,28 @@ def process_metric_update_namespace(namespace):
             None, 'incorrect usage: specify --api when hour or minute metrics are enabled')
 
 
+def validate_subnet(namespace):
+    from azure.cli.core.commands.arm import resource_id, is_valid_resource_id
+    from azure.cli.core.commands.client_factory import get_subscription_id
+
+    subnet = namespace.subnet
+    subnet_is_id = is_valid_resource_id(subnet)
+    vnet = namespace.vnet_name
+
+    if (subnet_is_id and not vnet) or (not subnet and not vnet):
+        return
+    elif subnet and not subnet_is_id and vnet:
+        namespace.subnet = resource_id(
+            subscription=get_subscription_id(),
+            resource_group=namespace.resource_group_name,
+            namespace='Microsoft.Network',
+            type='virtualNetworks',
+            name=vnet,
+            child_type='subnets',
+            child_name=subnet)
+    else:
+        raise CLIError('incorrect usage: [--subnet ID | --subnet NAME --vnet-name NAME]')
+
 # endregion
 
 # region TYPES
@@ -850,3 +879,42 @@ def get_char_options_validator(types, property_name):
 
         setattr(namespace, property_name, service_types)
     return _validator
+
+
+def page_blob_tier_validator(namespace):
+    if not namespace.tier:
+        return
+
+    if namespace.blob_type != 'page' and namespace.tier:
+        raise ValueError('Blob tier is only applicable to page blobs on premium storage accounts.')
+
+    try:
+        namespace.tier = getattr(get_sdk(ResourceType.DATA_STORAGE, 'blob.models#PremiumPageBlobTier'), namespace.tier)
+    except AttributeError:
+        from azure.cli.command_modules.storage.util import get_blob_tier_names
+        raise ValueError('Unknown premium page blob tier name. Choose among {}'.format(', '.join(
+            get_blob_tier_names('PremiumPageBlobTier'))))
+
+
+def block_blob_tier_validator(namespace):
+    if not namespace.tier:
+        return
+
+    if namespace.blob_type != 'block' and namespace.tier:
+        raise ValueError('Blob tier is only applicable to block blobs on standard storage accounts.')
+
+    try:
+        namespace.tier = getattr(get_sdk(ResourceType.DATA_STORAGE, 'blob.models#StandardBlobTier'), namespace.tier)
+    except AttributeError:
+        from azure.cli.command_modules.storage.util import get_blob_tier_names
+        raise ValueError('Unknown block blob tier name. Choose among {}'.format(', '.join(
+            get_blob_tier_names('StandardBlobTier'))))
+
+
+def blob_tier_validator(namespace):
+    if namespace.blob_type == 'page':
+        page_blob_tier_validator(namespace)
+    elif namespace.blob_type == 'block':
+        block_blob_tier_validator(namespace)
+    else:
+        raise ValueError('Blob tier is only applicable to block or page blob.')