+++ src/command_modules/azure-cli-appservice/HISTORY.rst
@@ -2,13 +2,15 @@
 
 Release History
 ===============
-Unreleased
-++++++++++++++++++
+0.1.14 (2017-08-28)
++++++++++++++++++++
 Breaking Change:webapp: fix inconsistencies in the output of "az webapp config appsettings delete/set"
 webapp: add a new alias of '-i' for "az webapp config container set --docker-custom-image-name"
 webapp: expose 'az webapp log show'
 webapp: expose new arguments from 'az webapp delete' to retain app service plan, metrics or dns registration. 
 webapp: detect a slot setting correctly 
+webapp: add param --docker-container-logging that goes through the same logic as --web-server-logging
+webapp: add premium v2 sku
 
 0.1.13 (2017-08-15)
 +++++++++++++++++++