+++ src/command_modules/azure-cli-resource/HISTORY.rst
@@ -13,7 +13,7 @@ unreleased
 +++++++++++++++++++
 * minor fixes
 
-unreleased
+2.0.13 (2017-08-28)
 +++++++++++++++++++
 * `group deployment create`: Fixes issue where templates which lacked "parameters" or "resources" failed to deploy.