+++ src/command_modules/azure-cli-storage/HISTORY.rst
@@ -5,9 +5,17 @@ Release History
 
 unreleased
 ++++++++++
+* Deduce content type and content encoding during blob and file upload if neither content type and content encoding are specified.
+
+2.0.13 (2017-08-28)
++++++++++++++++++++
+* Enable set blob tier
+* `storage account create/update`: Add `--bypass` and `--default-action` arguments to support service tunneling.
+* `storage account network-rule`: Added commands to add VNET rules and IP based rules.
 * Enable service encryption by customer managed key
 * Breaking change: rename --encryption option to --encryption-services for az storage account create and az storage account update command.
 * Fix #4220: az storage account update encryption - syntax mismatch
+
 
 2.0.12 (2017-08-11)
 +++++++++++++++++++